+++ Road Concept Backend/Road Concept Core/src/main/java/fr/enssat/lanniontech/core/managers/VehicleManager.java
@@ -65,7 +65,7 @@ public class VehicleManager {
         return true;
     }
 
-    public void newStep(double precision, boolean log, long timestamp) {
+    public void newStep(double precision, boolean log, int timestamp) {
         for (Vehicle activeVehicle : activeVehicles) {
             activeVehicle.updateAcceleration();
         }
