+++ Road Concept Backend/Road Concept Core/src/main/java/fr/enssat/lanniontech/core/roadElements/RoadMetrics.java
@@ -4,10 +4,10 @@ import java.util.UUID;
 
 public class RoadMetrics {
     private final UUID roadId;
-    private final long timestamp;
+    private final int timestamp;
     private final int congestion;
 
-    public RoadMetrics(UUID roadId, long timestamp, int congestion) {
+    public RoadMetrics(UUID roadId, int timestamp, int congestion) {
         this.roadId = roadId;
         this.timestamp = timestamp;
         this.congestion = congestion;
@@ -21,7 +21,7 @@ public class RoadMetrics {
         return congestion;
     }
 
-    public long getTimestamp() {
+    public int getTimestamp() {
         return timestamp;
     }
 }
